% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{std_error}
\alias{std_error}
\title{Standard error}
\usage{
std_error(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
The standard error of \code{x}.
}
\description{
Calculates the standard error of \code{x}
}
\examples{
x <- rnorm(10)
std_error(x)
}
